+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsChunkSource.java
@@ -481,6 +481,13 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     seenExpectedPlaylistError = false;
     expectedPlaylistUrl = null;
 
+    long chunkDurationUs = 0;
+    if (selectedTrackIndex < mediaChunkIterators.length
+        && mediaChunkIterators[selectedTrackIndex].next()) {
+      chunkDurationUs =
+          mediaChunkIterators[selectedTrackIndex].getChunkEndTimeUs()
+              - mediaChunkIterators[selectedTrackIndex].getChunkStartTimeUs();
+    }
     @Nullable
     CmcdLog cmcdLog =
         cmcdConfiguration == null
@@ -489,6 +496,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
                 cmcdConfiguration,
                 trackSelection,
                 bufferedDurationUs,
+                chunkDurationUs,
                 CmcdLog.STREAMING_FORMAT_HLS,
                 !playlist.hasEndTag);