+++ RELEASENOTES.md
@@ -34,6 +34,8 @@
         (([#33](https://github.com/androidx/media/issues/33)),([#9978](https://github.com/google/ExoPlayer/issues/9978))).
     *   Add fields streaming format (sf), stream type (st) and version (v) to
         Common Media Client Data (CMCD) logging.
+    *   Add fields top birate (tb) and object duration (d) to Common Media
+        Client Data (CMCD) logging.
 *   Transformer:
     *   Parse EXIF rotation data for image inputs.
     *   Remove `TransformationRequest.HdrMode` annotation type and its